refactor(chat): extract tool-call rendering into ToolCallContentPart
